@@ -1,74 +1,76 @@
 using DotnetSpider.AgentCenter.Store;
-using DotnetSpider.Statistics.Store;
+using DotnetSpider.Statistic.Store;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Design;
 using Microsoft.Extensions.Configuration;
 
 namespace DotnetSpider.Portal.Data
 {
-	public class PortalDbContext : DbContext, IDesignTimeDbContextFactory<PortalDbContext>
-	{
-		private readonly bool _isDesignTime;
+    public class PortalDbContext : DbContext, IDesignTimeDbContextFactory<PortalDbContext>
+    {
+        private readonly bool _isDesignTime;
 
-		public DbSet<Spider> Spiders { get; set; }
+        public DbSet<Spider> Spiders { get; set; }
 
-		public DbSet<SpiderHistory> SpiderHistories { get; set; }
+        public DbSet<SpiderHistory> SpiderHistories { get; set; }
 
-		public PortalDbContext()
-		{
-			_isDesignTime = true;
-		}
+        public PortalDbContext()
+        {
+            _isDesignTime = true;
+        }
 
-		public PortalDbContext(DbContextOptions<PortalDbContext> options, bool isDesignTime = false)
-			: base(options)
-		{
-			_isDesignTime = isDesignTime;
-		}
+        public PortalDbContext(DbContextOptions<PortalDbContext> options, bool isDesignTime = false)
+            : base(options)
+        {
+            _isDesignTime = isDesignTime;
+        }
 
-		protected override void OnModelCreating(ModelBuilder builder)
-		{
-			base.OnModelCreating(builder);
+        protected override void OnModelCreating(ModelBuilder builder)
+        {
+            base.OnModelCreating(builder);
 
-			if (!_isDesignTime)
-			{
-				builder.Model.AddEntityType(typeof(AgentInfo));
-				builder.Model.AddEntityType(typeof(AgentHeartbeat));
-				builder.Model.AddEntityType(typeof(SpiderStatistics));
-				builder.Model.AddEntityType(typeof(AgentStatistics));
-			}
+            if (!_isDesignTime)
+            {
+                builder.Model.AddEntityType(typeof(AgentInfo));
+                builder.Model.AddEntityType(typeof(AgentHeartbeat));
+                builder.Model.AddEntityType(typeof(SpiderStatistic));
+                builder.Model.AddEntityType(typeof(AgentStatistic));
+            }
 
-			builder.Entity<Spider>().HasIndex(x => x.Name);
-			builder.Entity<Spider>().HasIndex(x => x.CreationTime);
+            builder.Entity<Spider>().HasIndex(x => x.Name);
+            builder.Entity<Spider>().HasIndex(x => x.CreationTime);
 
-			builder.Entity<SpiderHistory>().HasIndex(x => x.Batch);
-			builder.Entity<SpiderHistory>().HasIndex(x => x.SpiderId);
-			builder.Entity<SpiderHistory>().HasIndex(x => x.CreationTime);
-		}
+            builder.Entity<SpiderHistory>().HasIndex(x => x.Batch);
+            builder.Entity<SpiderHistory>().HasIndex(x => x.SpiderId);
+            builder.Entity<SpiderHistory>().HasIndex(x => x.CreationTime);
+        }
 
-		public PortalDbContext CreateDbContext(string[] args)
-		{
-			var builder = new DbContextOptionsBuilder<PortalDbContext>();
+        public PortalDbContext CreateDbContext(string[] args)
+        {
+            var builder = new DbContextOptionsBuilder<PortalDbContext>();
 
-			var configurationBuilder = new ConfigurationBuilder();
-			configurationBuilder.AddJsonFile("appsettings.json");
-			var configuration = configurationBuilder.Build();
-			var options = new PortalOptions(configuration);
-			switch (options.DatabaseType?.ToLower())
-			{
-				case "mysql":
-					{
-						builder.UseMySql(ServerVersion.AutoDetect(options.ConnectionString));
-						break;
-					}
+            var configurationBuilder = new ConfigurationBuilder();
+            configurationBuilder
+                .AddJsonFile("appsettings.json")
+                .AddJsonFile("appsettings.migration.json", optional: true);
+            var configuration = configurationBuilder.Build();
+            var options = new PortalOptions(configuration);
+            switch (options.DatabaseType?.ToLower())
+            {
+                case "mysql":
+                    {
+                        builder.UseMySql(ServerVersion.AutoDetect(options.ConnectionString));
+                        break;
+                    }
 
-				default:
-					{
-						builder.UseSqlServer(options.ConnectionString);
-						break;
-					}
-			}
+                default:
+                    {
+                        builder.UseSqlServer(options.ConnectionString);
+                        break;
+                    }
+            }
 
-			return new PortalDbContext(builder.Options, true);
-		}
-	}
+            return new PortalDbContext(builder.Options, true);
+        }
+    }
 }
